+++ flexmeasures/data/schemas/scheduling/storage.py
@@ -27,6 +27,7 @@ from flexmeasures.data.schemas.sensors import (
     InflexibleDeviceSchema,
     OutputSensorReferenceSchema,
     VariableQuantityField,
+    PriceField,
 )
 from flexmeasures.utils.unit_utils import (
     ur,
@@ -520,6 +521,14 @@ class StorageFlexModelSchema(Schema):
         metadata=metadata.COUPLING_COEFFICIENT.to_dict(),
     )
 
+    soc_value_at_end = PriceField(
+        "/MWh",
+        data_key="soc-value-at-end",
+        required=False,
+        value_validator=validate.Range(min=0),
+        metadata=metadata.SOC_VALUE_AT_END.to_dict(),
+    )
+
     def __init__(
         self,
         start: datetime,
@@ -809,6 +818,13 @@ class DBStorageFlexModelSchema(Schema):
         required=False,
         validate=validate.Length(min=1),
         metadata={"deprecated field": "soc-usage"},
+    )
+
+    soc_value_at_end = PriceField(
+        "/MWh",
+        data_key="soc-value-at-end",
+        required=False,
+        value_validator=validate.Range(min=0),
     )
 
     roundtrip_efficiency = EfficiencyField(